+++ examples/persistent_state/lib/app/my_app.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_meedu/ui.dart';
 
 import 'ui/routes/app_routes.dart';
 import 'ui/routes/routes.dart';
@@ -10,12 +9,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      key: router.appKey,
       title: 'Hello world',
-      navigatorKey: router.navigatorKey,
-      navigatorObservers: [
-        router.observer,
-      ],
       routes: appRoutes,
       initialRoute: Routes.CITY_PICKER,
       theme: ThemeData(